Break the declaration of the static mapping into a file of its own.
