+++ include/deal.II/grid/grid_tools.h
@@ -28,7 +28,7 @@
 #  include <deal.II/dofs/dof_handler.h>
 
 #  include <deal.II/fe/mapping.h>
-#  include <deal.II/fe/mapping_q1.h>
+#  include <deal.II/fe/mapping_linear.h>
 
 #  include <deal.II/grid/manifold.h>
 #  include <deal.II/grid/tria.h>
