+++ include/deal.II/numerics/fe_field_function.h
@@ -27,7 +27,7 @@
 #include <deal.II/dofs/dof_accessor.h>
 #include <deal.II/dofs/dof_handler.h>
 
-#include <deal.II/fe/mapping_q1.h>
+#include <deal.II/fe/mapping_linear.h>
 
 #include <deal.II/grid/grid_tools_cache.h>
