+++ include/deal.II/numerics/matrix_creator.templates.h
@@ -29,7 +29,7 @@
 
 #include <deal.II/fe/fe.h>
 #include <deal.II/fe/fe_values.h>
-#include <deal.II/fe/mapping_q1.h>
+#include <deal.II/fe/mapping_linear.h>
 
 #include <deal.II/grid/tria_iterator.h>
