+++ include/deal.II/numerics/vector_tools_boundary.templates.h
@@ -26,6 +26,7 @@
 #include <deal.II/fe/fe_nothing.h>
 #include <deal.II/fe/fe_raviart_thomas.h>
 #include <deal.II/fe/fe_system.h>
+#include <deal.II/fe/mapping_linear.h>
 
 #include <deal.II/hp/fe_values.h>
 #include <deal.II/hp/q_collection.h>
