+++ include/deal.II/numerics/vector_tools_point_gradient.templates.h
@@ -18,6 +18,7 @@
 
 
 #include <deal.II/fe/fe_values.h>
+#include <deal.II/fe/mapping_linear.h>
 
 #include <deal.II/grid/grid_tools.h>
