+++ include/deal.II/numerics/vector_tools_project.templates.h
@@ -17,6 +17,8 @@
 #define dealii_vector_tools_project_templates_h
 
 
+#include <deal.II/fe/mapping_linear.h>
+
 #include <deal.II/lac/block_vector.h>
 #include <deal.II/lac/la_parallel_block_vector.h>
 #include <deal.II/lac/la_parallel_vector.h>
