+++ source/fe/mapping_linear.inst.in
@@ -17,6 +17,6 @@
 for (deal_II_dimension : DIMENSIONS; deal_II_space_dimension : SPACE_DIMENSIONS)
   {
 #if deal_II_dimension <= deal_II_space_dimension
-    template class MappingQ1<deal_II_dimension, deal_II_space_dimension>;
+    template struct StaticMappingQ1<deal_II_dimension, deal_II_space_dimension>;
 #endif
   }
