+++ source/grid/reference_cell.cc
@@ -16,7 +16,7 @@
 #include <deal.II/base/quadrature_lib.h>
 
 #include <deal.II/fe/mapping_fe.h>
-#include <deal.II/fe/mapping_q1.h>
+#include <deal.II/fe/mapping_linear.h>
 
 #include <deal.II/grid/grid_generator.h>
 #include <deal.II/grid/reference_cell.h>
